% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R, R/run_report.R
\name{generate_html_report}
\alias{generate_html_report}
\title{A function to generate a html report}
\usage{
generate_html_report(
  full_path_2_Rdatafile = "GeneratedData.Rdata",
  dir_4_report = ".",
  path_2_Rmd_template = system.file("rmarkdown", package = "exrender")
)

generate_html_report(
  full_path_2_Rdatafile = "GeneratedData.Rdata",
  dir_4_report = ".",
  path_2_Rmd_template = system.file("rmarkdown", package = "exrender")
)
}
\arguments{
\item{full_path_2_Rdatafile}{full path to the Rdatafile}

\item{dir_4_report}{directory to place the report}

\item{path_2_Rmd_template}{full path to the html report template}
}
\value{
Writes a html report to file

Writes a html report to file
}
\description{
This function generates the html report.

This function generates the html report.
}
\examples{
generate_html_report()
generate_html_report()
}
\keyword{knit}
\keyword{html}
\keyword{report}
