% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R, R/run_report.R
\name{generate_report}
\alias{generate_report}
\title{A function to generate the Rmd report}
\usage{
generate_report(
  path_2_Rdatafile = ".",
  dir_4_report = ".",
  path_2_Rmd_template = system.file("rmarkdown", package = "exrender")
)

generate_report(
  path_2_Rdatafile = ".",
  dir_4_report = ".",
  path_2_Rmd_template = system.file("rmarkdown", package = "exrender")
)
}
\arguments{
\item{path_2_Rdatafile}{full path to the Rdatafile}

\item{dir_4_report}{directory to place the report}

\item{path_2_Rmd_template}{full path to the Rmd report template}
}
\value{
Writes a pdf report to file

Writes a pdf report to file
}
\description{
This function generates the Rmd report.

This function generates the Rmd report.
}
\examples{
generate_report()
generate_report()
}
\keyword{pdf}
\keyword{render}
\keyword{report}
