% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_some_data.R
\name{make_some_data}
\alias{make_some_data}
\title{A function to make some data}
\usage{
make_some_data()
}
\value{
writes a data set "wdata" a prcomp object "my_pca" and a ggplot object "my_plot" to a Rdata file.
}
\description{
This function generates some random data to pass to my Markdown Report
}
\examples{
make_some_data()
}
\keyword{data}
\keyword{generation}
\keyword{random}
